style: pet doxygen and clang-tidy

+++ src/triangulate/triangulate2DZ.cpp
@@ -62,7 +62,7 @@ triangulate2DZ(const Triangle                  &g,
 
   for (size_t i = 0; i < 4; i++) {
     Vertex_handle const vertex =
-        triangulation.addVertex(g.vertex(i).coordinate());
+        triangulation.addVertex(g.vertex(static_cast<int>(i)).coordinate());
 
     if (i != 0) {
       triangulation.addConstraint(last, vertex);
@@ -79,12 +79,6 @@ triangulate2DZ(const Triangle                  &g,
 // ----------------------------------------------------------------------------------
 /// @publicsection
 
-/**
- * @brief Triangulate a geometry collection in 2D with Z values
- *
- * @param g The geometry collection to triangulate
- * @param triangulation The triangulation to add vertices and constraints to
- */
 void
 triangulateCollection2DZ(const Geometry                  &g,
                          ConstraintDelaunayTriangulation &triangulation)
@@ -94,12 +88,6 @@ triangulateCollection2DZ(const Geometry                  &g,
   }
 }
 
-/**
- * @brief Triangulate a geometry in 2D with Z values
- *
- * @param g The geometry to triangulate
- * @param triangulation The triangulation to add vertices and constraints to
- */
 void
 triangulate2DZ(const Geometry                  &g,
                ConstraintDelaunayTriangulation &triangulation)
@@ -160,13 +148,7 @@ triangulate2DZ(const Geometry                  &g,
   }
 }
 
-/**
- * @brief Triangulate a geometry in 2D with Z values and return the
- * triangulation
- *
- * @param g The geometry to triangulate
- * @return The constraint Delaunay triangulation
- */
+/// @private
 auto
 triangulate2DZ(const Geometry &g) -> ConstraintDelaunayTriangulation
 {
